% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lre.R
\name{bk_condition}
\alias{bk_condition}
\title{Check Blanchard-Khan condition}
\usage{
bk_condition(nx, ns, silent = TRUE)
}
\arguments{
\item{nx}{Number of predetermined variables.}

\item{ns}{Number of stable eigenvalues.}

\item{silent}{Boolean. If set FALSE, violation of BK condition causes an exception.}
}
\value{
Boolean. TRUE if nx = ns.
}
\description{
Check Blanchard-Khan condition
}
