% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbind.R
\name{dbind}
\alias{dbind}
\title{Diagonal bind for a matrix, source: Patrick Blisle, http://www.medicine.mcgill.ca/epidemiology/joseph/pbelisle/dbind.html}
\usage{
dbind(m, a, rev = F)
}
\arguments{
\item{m}{Upper left matrix(or element)}

\item{a}{Bottom right matrix}

\item{rev}{reverces the order if necessary}
}
\value{
A new matrix with the two inputs in its block diagonal.
}
\description{
Diagonal bind for a matrix, source: Patrick Blisle, http://www.medicine.mcgill.ca/epidemiology/joseph/pbelisle/dbind.html
}
