% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein.R
\name{g}
\alias{g}
\title{a function that computes the non-predetermined from the predetermined}
\usage{
g(x1, coefficients, y)
}
\arguments{
\item{x1}{the predetermied variable}

\item{coefficients}{coeffients matrixes that is provid by klein_coefficients_from()}

\item{y}{an valu that from an array computed by yu()}
}
\description{
a function that computes the non-predetermined from the predetermined
}
