% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein.R
\name{h}
\alias{h}
\title{h a function that computes the predetermined of the next period}
\usage{
h(x1, u, coefficients, y)
}
\arguments{
\item{x1}{the predetermied variable}

\item{u}{shocks}

\item{coefficients}{coeffients matrixes that is provid by klein_coefficients_from()}

\item{y}{an valu that from an array computed by yu()}
}
\description{
h a function that computes the predetermined of the next period
}
