% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impluse}
\alias{impluse}
\title{impluse}
\usage{
impluse(g, h, x0, t, e1)
}
\arguments{
\item{g}{function g}

\item{h}{function h}

\item{x0}{steady state}

\item{t}{integer, simulation length}

\item{e1}{i-th element e[i] corresponds to a shock to i-th endogenous variable}
}
\value{
g and h functions
}
\description{
impluse
}
