% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse}
\alias{impulse}
\title{impulse
\code{}
Impulse response given (g,h,ξ1)}
\usage{
impulse(g, h, x0, t, e1)
}
\arguments{
\item{g}{function g}

\item{h}{function h}

\item{x0}{steady state. For linear models, it’s the origin.}

\item{t}{integer, simulation length}

\item{e1}{i-th element e[i] corresponds to a shock to i-th endogenous variable}
}
\value{
out: matrix of simulation output
}
\description{
impulse
\code{}
Impulse response given (g,h,ξ1)
}
