% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein.R
\name{klein_coefficients_from}
\alias{klein_coefficients_from}
\title{Making parameters accrding Klein(200)}
\usage{
klein_coefficients_from(decom)
}
\arguments{
\item{decom}{a list object made by qz_klein()}
}
\description{
Making parameters accrding Klein(200)
}
