% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lre_ar.R
\name{lre_ar}
\alias{lre_ar}
\title{lre_ar}
\usage{
lre_ar(A, E, B, Phi, nx)
}
\arguments{
\item{A}{matrix A}

\item{E}{matrix B}

\item{B}{matrix B}

\item{Phi}{matrix Φ}

\item{nx}{number od predetermined variables}
}
\value{
list(g = g, h = h):list of two functions(g,h)
}
\description{
\code{}
Compute (g,h) from (A,E,B,Φ,n1)
}
