% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrem.R
\name{lre_auto}
\alias{lre_auto}
\title{lrem}
\usage{
lre_auto(A, E, x0)
}
\arguments{
\item{A}{coeffients of previous period}

\item{E}{coeffients of latter period}

\item{x0}{double < 0, initial values}
}
\value{
g and h functions
}
\description{
lrem
}
