% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-bk.R
\name{lre_auto_bk}
\alias{lre_auto_bk}
\title{LRE solution method based on Blanchard and Kahn (1980, ECTA)}
\usage{
lre_auto_bk(A, nx)
}
\arguments{
\item{A}{Square matrix}

\item{nx}{The number of predetermined variables, \code{nx} is required
by the algorithm.}
}
\value{
List of two functions (g, h), passed to \code{\link{simulate}}
}
\description{
This function solves for a linear policy function for the Linear Rational
Expectations model of \deqn{(x_{t+1}, y_{t+1}) = A (x_{t}, y_{t})}{
(x_{t+1}, y_{t+1}) = A (x_{t}, y_{t})}, where x and y are predetermined and
non-predetermined variables, respectively.
}
