% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-klein.R
\name{lre_auto_klein}
\alias{lre_auto_klein}
\title{LRE solution method based on Klein (2000, JEDC)}
\usage{
lre_auto_klein(A, E, nx)
}
\arguments{
\item{A, E}{Square matrices of same size. \deqn{zE - A}{zE-A} is a regular
pencil}

\item{nx}{The number of predetermined variables, \code{nx} is required
by the algorithm.}
}
\value{
List of two functions (g, h), passed to \code{\link{simulate}}
}
\description{
This function solves for a linear policy function for the Linear Rational
Expectations model of \deqn{E (x_{t+1}, y_{t+1}) = A (x_{t}, y_{t})}{E
(x_{t+1}, y_{t+1}) = A (x_{t}, y_{t})}, where x and y are predetermined and
non-predetermined variables, respectively.
}
\references{
Klein (2000, JEDC) \link{https://doi.org/10.1016/S0165-1889(99)00045-7}
}
