% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-klein.R
\name{make_blocks_of}
\alias{make_blocks_of}
\title{making block matrix of a square matrix A}
\usage{
make_blocks_of(A)
}
\arguments{
\item{A}{a square matrix A}
}
\description{
making block matrix of a square matrix A
}
