% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified-determinant.R
\name{modified_det}
\alias{modified_det}
\title{Obtain determinant even when input is scalar}
\usage{
modified_det(A)
}
\arguments{
\item{A}{there exists n in N such that length(A) = n double, n_by_n matrix or scalar}
}
\value{
Determinant if A is matrix, otherwise; itself
}
\description{
Obtain determinant even when input is scalar
}
