% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schur-decomposition-with-proper-ordering.R
\name{proper_order_sch}
\alias{proper_order_sch}
\title{Schur decomposition with stable eigenvalues in the upper-left block)}
\usage{
proper_order_sch(A)
}
\arguments{
\item{A}{dim(A[1]) = dim(A)[2], n_by_n matrix}
}
\value{
Schur decomposition of this matrix with stable eigenvalues in the upper-left block
}
\description{
Schur decomposition with stable eigenvalues in the upper-left block)
}
