% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{qz}
\alias{qz}
\title{QZ decomposition}
\usage{
qz(A, B)
}
\arguments{
\item{A, B}{Sqaure matrices of same size. (A, B) is a regular pencil.}
}
\value{
QZ decomposition.
}
\description{
Real QZ decomposition. Weakly stable generalized eigenvalues (in
discrete-time) are collected at the upper-left corner.
}
