% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein.R
\name{qz_klein}
\alias{qz_klein}
\title{QZ decompose of pair matrixes A and E}
\usage{
qz_klein(A, E)
}
\arguments{
\item{A}{a legurar matrix that is the same size as E}

\item{E}{a legurar matrix that is the same size as E}
}
\description{
QZ decompose of pair matrixes A and E
}
