% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-klein.R
\name{reverse_1dim_array}
\alias{reverse_1dim_array}
\title{returns reversed array of original array}
\usage{
reverse_1dim_array(original)
}
\arguments{
\item{original}{an array to be reversed}
}
\description{
returns reversed array of original array
}
