% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{schur}
\alias{schur}
\title{Schur decomposition}
\usage{
schur(A)
}
\arguments{
\item{A}{Sqaure matrix.}
}
\value{
Schur decomposition of \code{A}.
}
\description{
Real Schur decomposition. Weakly stable eigenvalues (in discrete-time) are
collected at the upper-left corner.
}
