% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{simulate
\code{}
Simulation given (g,h,x^1_0,ξ)}
\usage{
simulate(g, h, x0, t, e)
}
\arguments{
\item{g}{function g}

\item{h}{function h}

\item{x0}{initial x^1_0}

\item{t}{integer, simulation length}

\item{e}{vector or matrix, each row e[k, ] corresponds to ξ_k+1}
}
\value{
out: matrix of simulation output
}
\description{
simulate
\code{}
Simulation given (g,h,x^1_0,ξ)
}
