% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_shock}
\alias{simulate_shock}
\title{Simulating an LRE with an exogenous shock}
\usage{
simulate_shock(g, h, x0, t, e)
}
\arguments{
\item{g}{Decision rule}

\item{h}{Motion rule}

\item{x0}{vector of predetermined variables that includes the AR parameter}

\item{t}{length of the simulation}

\item{e}{vector of exogenous shocks, input is a list which is then transformed into an appropriate matrix}
}
\value{
A matrix showing the dynamic path of variables over the simulation time
}
\description{
Simulating an LRE with an exogenous shock
}
