% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{impulse}
\alias{impulse}
\title{Compute impulse response}
\usage{
impulse(g, h, x0, t, e1)
}
\arguments{
\item{g}{maps predetermined to non-predetermined}

\item{h}{maps predetermined forward in time}

\item{x0}{initial values for predetermined variables. For linear case, they
are zero.}

\item{t}{integer, length of simulation}

\item{e1}{impulse shock}
}
\value{
matrix comprising of simulation output
}
\description{
Compute impulse response
}
\seealso{
\code{\link{simulate}}
}
