% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lre.R
\name{lre_ar}
\alias{lre_ar}
\title{LRE solution method under AR inputs}
\usage{
lre_ar(A, E = NULL, B, Phi, nx = NULL, ny = NULL, x0 = NULL)
}
\arguments{
\item{A, E}{Square matrices of same size. \deqn{zE - A}{zE-A} is a regular
pencil. E = NULL is understood as E = I.}

\item{B}{coefficient matrix for input term}

\item{Phi}{matrix. Inputs are assumed to follow VAR process u_{t+1} = Phi %*%
u_{t} + epsilon_{t+1}}

\item{nx, ny, x0}{The number of predetermined variables, \code{nx}, is required
by the algorithm. It is passed by \code{nx} or computed from the number of
the non-predetermined \code{ny}, or the initial vector \code{x0}}
}
\value{
List of two functions (g, h), passed to \code{\link{simulate}}
}
\description{
This function solves for a linear policy function for the Linear Rational
Expectations model of \deqn{E (x_{t+1}, y_{t+1}) = A (x_{t}, y_{t}) +
Bu_{t}}{E (x_{t+1}, y_{t+1}) = A (x_{t}, y_{t}) + Bu_{t}}, where x and y are
predetermined and non-predetermined variables, respectively. The input term
follows AR process.
}
