% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrem.R
\name{lre_ar}
\alias{lre_ar}
\title{Solving LRE with exogenous shocks of the AR variety}
\usage{
lre_ar(A, E, B, Phi, n)
}
\arguments{
\item{A}{Coefficient matrix on x(t)}

\item{E}{Coefficient matrix on x(t+1)}

\item{B}{Coefficient matrix on the error term u}

\item{Phi}{The AR process for the error term}

\item{n}{number of predetermined variables(excluding the exogenous shocks)}
}
\value{
Motion and decision rule functions for the LRE with AR exogenous shocks
}
\description{
Solving LRE with exogenous shocks of the AR variety
}
