% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lre_ar.R
\name{lre_ar}
\alias{lre_ar}
\title{lre_ar}
\usage{
lre_ar(A, E, B, Phi, nx)
}
\arguments{
\item{A}{coeffients of previous period}

\item{E}{coeffients of latter period}

\item{B}{matrixB}

\item{Phi}{matrixPhi}

\item{nx}{number od predetermined variables}
}
\value{
g and h functions
}
\description{
lre_ar
}
