% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lre_ar.R
\name{lre_ar}
\alias{lre_ar}
\title{Solve a linear rational expectation (LRE) system with autoregressive (AR) inputs}
\usage{
lre_ar(A, E, B, Phi, x0, u0)
}
\arguments{
\item{A}{dim(A)[1] = dim(A)[2], n_by_n matrix}

\item{E}{dim(E)[1] = dim(E)[2], n_by_n matrix}

\item{B}{dim(B)[1] = or (neq) dim(B)[2], n_by_m matrix}

\item{Phi}{dim(Phi)[1] = dim(Phi)[2], m_by_m matrix}

\item{x0}{x0 in R, Initial value of predetermined vector including inital shock}

\item{u0}{u0 in R, Initial value of AR inputs (vector or scalar)}
}
\value{
Functions of predetermined vector at t on un-predetermined vector at t and predetermined vector at t+1
}
\description{
Solve a linear rational expectation (LRE) system with autoregressive (AR) inputs
}
