% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lre.R
\name{lre_auto}
\alias{lre_auto}
\title{LRE solution method}
\usage{
lre_auto(A, E = NULL, nx = NULL, ny = NULL, x0 = NULL)
}
\arguments{
\item{A, E}{Square matrices of same size. \deqn{zE - A}{zE-A} is a regular
pencil. E = NULL is understood as E = I.}

\item{nx, ny, x0}{The number of predetermined variables, \code{nx}, is required
by the algorithm. It is passed by \code{nx} or computed from the number of the
non-predetermined \code{ny}, or the initial vector \code{x0}}
}
\value{
List of two functions (g, h), passed to \code{\link{simulate}}
}
\description{
This function solves for a linear policy function for the Linear Rational
Expectations model of \deqn{E (x_{t+1}, y_{t+1}) = A (x_{t}, y_{t})}{E
(x_{t+1}, y_{t+1}) = A (x_{t}, y_{t})}, where x and y are predetermined and
non-predetermined variables, respectively. When E is NULL, it is understood
as E = I (identity matrix) and Schur decomposition of A is used to compute
the solution. When E is not NULL, the generalized Schur decomposition of (A,
E) is used.
}
\references{
Klein (2000, JEDC)
  \link{https://doi.org/10.1016/S0165-1889(99)00045-7}
}
\seealso{
\code{\link{lre_auto_klein}}, \code{\link{lre_auto_bk}}
}
