% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lre_auto.R
\name{lre_auto}
\alias{lre_auto}
\title{lre_auto}
\usage{
lre_auto(A, E, nx)
}
\arguments{
\item{A}{coeffients of previous period}

\item{E}{coeffients of latter period}

\item{nx}{number od predetermined variables}
}
\value{
g and h functions
}
\description{
lre_auto
}
