% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lre_auto_bk.R
\name{lre_auto_bk}
\alias{lre_auto_bk}
\title{Solve a linear autonomous system}
\usage{
lre_auto_bk(A, x0)
}
\arguments{
\item{A}{dim(A[1]) = dim(A)[2], n_by_n matrix, n_by_n matrix}

\item{x0}{x0 in R, Initial value of predetermined vector}
}
\value{
Functions of predetermined vector at t on un-predetermined vector at t and predetermined vector at t+1
}
\description{
Solve a linear autonomous system
}
