% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{simulate_ssr_no_output}
\alias{simulate_ssr}
\title{Simulate state space model}
\usage{
simulate(g = NULL, h, x0, t)

simulate_ssr_no_output(h, x0, t)

simulate_ssr(g, h, x0, t)
}
\arguments{
\item{g}{maps predetermined to non-predetermined}

\item{h}{maps predetermined forward in time}

\item{x0}{initial values for predetermined variables}

\item{t}{length of simulation}
}
\value{
matrix comprising of simulation output
}
\description{
This function simulates a state space model with a state equation,
x_{t+1} = h(x_{t}), and output equation, y_{t} = g(x_{t}), from initial condition
x_{0}. In macroeconomics, x is said to be the vector of predetermined variables and
y the vector of non-predetermined variables.
}
