% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein.R
\name{simulate}
\alias{simulate}
\title{simulate a model using function that decides pre and non-pre determinant variables}
\usage{
simulate(A, E, g, h, x0, t, u)
}
\arguments{
\item{A}{a legular matrix that is the same size as E}

\item{E}{a legurar matrix that is the same size as E}

\item{g}{a function that computes the non-predetermined from the predetermined}

\item{h}{a function that computes the predetermined of the next period}

\item{x0}{an initial condition}

\item{t}{a number of times to caluculate}

\item{u}{an array that contains shocks}
}
\description{
simulate a model using function that decides pre and non-pre determinant variables
}
