% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate}
\alias{simulate}
\title{Wrapper function for the simulations.}
\usage{
simulate(g, h, x0, t = NULL, e = NULL)
}
\arguments{
\item{g}{First output of a given lre function}

\item{h}{Second output of the lre function}

\item{x0}{vector of initial predetermined variables. If dealing with an AR process, this needs to include the Phi elements}

\item{t}{length of the simulation. If not given, assumed to be the length of the exogenous shocks + 1}

\item{e}{exogenous shocks}
}
\value{
A matrix showing the dynamic paths of the variables in the specification
}
\description{
Wrapper function for the simulations.
}
