% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(g, h, x0, t, e)
}
\arguments{
\item{g}{function g}

\item{h}{function h}

\item{x0}{steady state}

\item{t}{integer, simulation length}

\item{e}{vector or matrix, each row e[k, ]}
}
\value{
g and h functions
}
\description{
simulate
}
