% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-on-linear-system.R
\name{simulate}
\alias{simulate}
\title{Simulate a linear system with given transition}
\usage{
simulate(g, h, x0, u0 = NULL, t)
}
\arguments{
\item{g}{function, Function of predetermined vector at t on un-predetermined vector at t}

\item{h}{function, Function of predetermined vector at t on predetermined vector at t+1}

\item{x0}{x0 in R, Initial value of predetermined vector}

\item{t}{t in N, Length of simulation}

\item{e}{vector or matrix, each row e[k, ] corresponds to shock at t = k}
}
\value{
Path of all vectors
}
\description{
Simulate a linear system with given transition
}
